Refactor photo storage
The photos are now stored in folder with the user_id of the user who upload the image - this is the ensure no exposer of photos even when the crc32 is the same for different users.
The thumbnail is stored in storage instead of DataStore.
The photos API support passing photo label which currently support only thumbnail or main.